requirements.txt
Django==1.10.2
Pillow==3.3.1
django-debug-toolbar==1.6
#mysqlclient==1.3.7
